requirements.txt
aiofiles
bcrypt==4.0.1
fastapi
fastapi[standard]
passlib==1.7.4
python-jose
python-multipart
SQLAlchemy
uvicorn
python-dotenv
psycopg2
Jinja2
alembic
python-dotenv
